# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management API",
        "description": "API documentation for the Event Management System. This system allows managing yearly events with all their associated content including summaries, themes, programmes, speakers, sponsors, FAQs, media, galleries, and attendances.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/events/{eventId}/summaries": {
            "post": {
                "tags": ["Event Content"],
                "summary": "Add summary to event",
                "operationId": "addSummary",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["summary"],
                                "properties": {
                                    "summary": {"type": "string", "example": "This is the event description..."}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Summary created successfully"
                    }
                }
            }
        },
        "/api/events/{eventId}/themes": {
            "post": {
                "tags": ["Event Content"],
                "summary": "Add theme to event",
                "operationId": "addTheme",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["theme"],
                                "properties": {
                                    "theme": {"type": "string", "example": "AI & Machine Learning"},
                                    "description": {"type": "string", "example": "Exploring AI technologies"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Theme created successfully"
                    }
                }
            }
        },
        "/api/events/{eventId}/speakers": {
            "post": {
                "tags": ["Event Content"],
                "summary": "Add speaker to event",
                "operationId": "addSpeaker",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name"],
                                "properties": {
                                    "name": {"type": "string", "example": "Dr. Jane Smith"},
                                    "title": {"type": "string", "example": "CTO"},
                                    "organization": {"type": "string", "example": "Tech Corp"},
                                    "bio": {"type": "string", "example": "Expert in AI"},
                                    "email": {"type": "string", "format": "email"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Speaker created successfully"
                    }
                }
            }
        },
        "/api/events/{eventId}/sponsors": {
            "post": {
                "tags": ["Event Content"],
                "summary": "Add sponsor to event",
                "operationId": "addSponsor",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name"],
                                "properties": {
                                    "name": {"type": "string", "example": "Tech Corporation"},
                                    "tier": {"type": "string", "example": "Gold"},
                                    "website": {"type": "string", "example": "https://techcorp.com"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sponsor created successfully"
                    }
                }
            }
        },
...
    },
...
}